maj auto

+++ TestBS/Sensors.cpp
@@ -20,7 +20,7 @@ bool Sensors::getTouch(int button)
 {
 	return 0;
 }
-float Sensors::getSonar(uint8_t num) // valeur du sonar
+int Sensors::getSonar(uint8_t num) // valeur du sonar
 {
 	//BP.get_sensor(num,&sonar1);
 	//myi2c.address = 0x02;
@@ -28,7 +28,7 @@ float Sensors::getSonar(uint8_t num) // valeur du sonar
 	//myi2c.length_read = 1;
 	printf("result = %d\n",BP.transact_i2c(num, &myi2c));
 	
-	return (float) (myi2c.buffer_read[0]);//+myi2c.buffer_read[1]+myi2c.buffer_read[2]);
+	return (int) (myi2c.buffer_read[0]);//+myi2c.buffer_read[1]+myi2c.buffer_read[2]);
 	//return sonar1.cm;
 }
 int Sensors::readEncoder(void)
