+++ TestBS/Sensors.h
@@ -5,7 +5,7 @@ public:
 Sensors(void); // constructeur
 bool getTouch(int button); // on récupè-re la valeur du contacteur en booléen
 int readEncoder(void); //lance la lecture des codeurs
-float getSonar(uint8_t num); // valeur du sonar
+int getSonar(uint8_t num); // valeur du sonar
 int getEncoder(int *right,int *left); //
 };
